% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatMetRegulon.R
\name{CalEnzymeAct}
\alias{CalEnzymeAct}
\title{CalEnzymeAct
Calculation of metabolic enzyme activities using regulatory networks}
\usage{
CalEnzymeAct(regulon, eset, ...)
}
\arguments{
\item{regulon}{Regulon object}

\item{eset}{a gene expression matrix with genes (probes) in rows and samples in columns}

\item{...}{Parameters passed to viper}
}
\value{
A matrix of inferred enzyme activity for each regulator gene in the network across all samples
}
\description{
CalEnzymeAct
Calculation of metabolic enzyme activities using regulatory networks
}
\examples{
data(testExp)
data(network)
regulon <- CreatMetRegulon(network, as.matrix(testExp))
enzyme_activity <- CalEnzymeAct(regulon,as.matrix(testExp))
}
