% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatMetRegulon.R
\name{CreatMetRegulon}
\alias{CreatMetRegulon}
\title{CreatMetRegulon
Create regulon object based on ARACNe-AP results}
\usage{
CreatMetRegulon(afile, eset)
}
\arguments{
\item{afile}{Character string indicating the name of the ARACNe network file}

\item{eset}{a gene expression matrix with genes (probes) in rows and samples in columns}
}
\value{
Regulon object
}
\description{
CreatMetRegulon
Create regulon object based on ARACNe-AP results
}
\examples{
data(testExp)
data(network)
regulon <- CreatMetRegulon(network, as.matrix(testExp))
}
