% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MetabolicEnzymes}
\alias{MetabolicEnzymes}
\title{MetabolicEnzymes}
\format{
A character vector
}
\usage{
MetabolicEnzymes
}
\description{
A list of known metabolic enzyme genes.
}
\keyword{datasets}
