% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Metabolism_Re_genes}
\alias{Metabolism_Re_genes}
\title{Metabolism_Re_genes}
\format{
A character vector
}
\usage{
Metabolism_Re_genes
}
\description{
Pre-selected metabolism-related genes identified through network propagation.
}
\keyword{datasets}
