% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metabolites}
\alias{metabolites}
\title{metabolites}
\format{
A character vector
}
\usage{
metabolites
}
\description{
A vector of metabolite molecule names used as seed nodes for the
Random Walk with Restart (RWR) algorithm.
}
\keyword{datasets}
