% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MetNetwork}
\alias{MetNetwork}
\title{MetNetwork}
\format{
A data.frame with two columns representing directed interactions:
\describe{
  \item{source}{source node}
  \item{target}{target node}
}
}
\usage{
MetNetwork
}
\description{
A metabolism-related biological network integrating protein-protein interactions,
metabolic enzyme-substrate, enzyme-metabolite, metabolite-receptor, and
metabolite-transporter interactions from KEGG, Reactome, Human-GEM, BRENDA,
TRRUST2, and STRING databases.
}
\keyword{datasets}
