% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regulon}
\alias{regulon}
\title{regulon}
\format{
A viper regulon object (list of transcriptional regulatory modules)
}
\usage{
regulon
}
\description{
A pre-computed regulon object derived from TCGA expression data using
ARACNe-AP network inference, used by CalEnzymeActWithTCGAregulon().
}
\keyword{datasets}
