% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testExp}
\alias{testExp}
\title{testExp}
\format{
A numeric matrix with genes in rows and samples in columns
}
\usage{
testExp
}
\description{
An example gene expression matrix used for testing and demonstration.
}
\keyword{datasets}
