% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatMetFegulon.R
\name{CalEnzymeAct}
\alias{CalEnzymeAct}
\title{CalEnzymeAct
Calculation of metabolic enzyme activities using regulatory networks}
\usage{
CalEnzymeAct(regulon, eset, ...)
}
\arguments{
\item{regulon}{Regulon object}

\item{eset}{a gene expression matrix with genes (probes) in rows and samples in columns}

\item{...}{Parameters passed to viper}
}
\value{
A matrix of inferred enzyme activity for each regulator gene in the network across all samples
}
\description{
CalEnzymeAct
Calculation of metabolic enzyme activities using regulatory networks
}
