% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalEnzymeActWithTCGAregulon.R
\name{CalEnzymeActWithTCGAregulon}
\alias{CalEnzymeActWithTCGAregulon}
\title{CalEnzymeActWithTCGAregulon}
\usage{
CalEnzymeActWithTCGAregulon(eset, TCGAtype = "HCC", ...)
}
\arguments{
\item{eset}{a gene expression matrix with genes (probes) in rows and samples in columns}

\item{TCGAtype}{Character string indicating cancer type(Def:HCC). Can be one of the following: "ACC","BLCA","BRCA","CESC","CHOL","COAD","DLBC","ESCA","GBM","HNSC","KICH","KIRC","KIRP","LAML","LGG","LIHC","LUAD","LUSC","MESO","OV","PAAD","PCPG","PRAD","READ","SARC","SKCM","STAD","TGCT","THCA","THYM","UCEC","UCS","UVM"}

\item{...}{... Parameters passed to viper}
}
\description{
CalEnzymeActWithTCGAregulon
}
