% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetGenes.R
\name{getMetGenes}
\alias{getMetGenes}
\title{getMetGenes
Obtaining metabolism-related genes on the metabolic biological network by restarting the walk random algorithm}
\usage{
getMetGenes(
  network,
  metabolites,
  filter.pct = 0.2,
  gamma = 0.7,
  tmax = 1000,
  eps = 1e-10,
  norm = TRUE
)
}
\arguments{
\item{network}{a two-column data.frame representing metabolism-related biological network. This Networkincludes protein-protein interactions, metabolic enzymes-metabolic substrates, metabolic enzymes-metabolites, metabolites-metabolite receptors, metabolites-transporters}

\item{metabolites}{a vector of strings representing metabolite molecules in a metabolism-related biological network}

\item{filter.pct}{a filter threshold (def: 0.2). Filter out genes not related to metabolism.}

\item{gamma}{restart parameter (def: 0.7)}

\item{tmax}{maximum number of iterations (steps) (def: 1000)}

\item{eps}{maximum allowed difference between the computed probabilities at the steady state (def. 1e-10)}

\item{norm}{whether to normalize the adjacency matrix (def: TRUE)}
}
\value{
a vector
}
\description{
getMetGenes
Obtaining metabolism-related genes on the metabolic biological network by restarting the walk random algorithm
}
\examples{
\dontrun{
data(MetNetwork)
data(metabolites)
getMetGenes(network,
metabolites,
filter.pct=0.2,
gamma = 0.7,
tmax = 1000,
eps = 1e-10,
norm = TRUE)
}
}
