% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreatMetFegulon.R
\name{readAracne}
\alias{readAracne}
\title{readAracne
read ARACNe-AP results}
\usage{
readAracne(afile, format = c("3col", "adj"))
}
\arguments{
\item{afile}{Character string indicating the name of the ARACNe network file}

\item{format}{Character string, indicating the format of the aracne file, either adj for adjacency matrixes generated by aracne, or 3col when the interactome is represented by a 3 columns text file, with regulator in the first column, target in the second and mutual information in the third column}
}
\value{
Regulon object
}
\description{
readAracne
read ARACNe-AP results
}
\examples{
regulon <- CreatMetFegulon('ARACNeOutputFile.txt',
}
